Implement a proper system for rate-limiting
When commands are sent to IRCClient#send, you can specify an optional
set of received numeric commands to wait for.
For example, when requesting NAMES, you can tell IRCClient to wait for 366
(names complete).

All such commands are queued up, and the next sent only when the response
to the previous has been received.